Extract to spec/support helper file
ident
